Implement the stop mode for the `stop command.
It is now possible to specify to the `stop` command a second argument
that is one of the supported modes for pg_ctl:
- smart 8the default)
- immediate
- fast
Abbreviations are also recognized (it is possible to specify only the first
letter of the command). If no argument is specified, smart mode is
applied.

Documentation updated.

Close #12