Don't use 'smart' as default stop mode.
See discussion
<https://github.com/theory/pgenv/issues/12#issuecomment-419094833>
In order to allow `pg_ctl` to use its own default stop mode, do not
use 'smart' when the user does not specify a stop mode. To achieve this,
a special variable PGENV_STOP_MODE handles the whole flag for `pg_ctl`
or nothing (default).

The variable PGENV_STOP_MODE can be passed to the script to set the
stop mode, so the following are same invocations:
   $ pgenv stop fast
   $ PGENV_STOP_MODE="fast" pgenv stop

The command line argument takes the precedence on the environment variable.
The environment variable does accept only the 'imemdiate, fast, smart' words
and nothing else, so it cannot be used to inject flags into pg_ctl.

Documentation updated.