[#25] Provide a way to build the current development version
The application already has a way to build from a local git
repository.
The idea of this commit is to provide a way to specify the HEAD
version to indicate PostgreSQL has to be built from the current git
checkout.
Specify the `git_tag` as a possible argument to `build git`.

Update the documentation accordingly.

See discussion and initial work in #26